jenkins upload용

+++ src/main/java/com/cafe24/jblog/repository/BlogDAO.java
@@ -57,8 +57,8 @@ public class BlogDAO {
 		return result;
 	}
 	
-	public int getFirstCategoryNoById(String id) {
-		int result = sqlSession.selectOne("getFirstCategoryNoById", id);
+	public Long getFirstCategoryNoById(String id) {
+		Long result = sqlSession.selectOne("getFirstCategoryNoById", id);
 		return result;
 	}
